moved name_cache into db
